Refactores wiki.examples to examples and fixes sample tests
