+++ org.jboss.reddeer.examples/build.properties
