+++ org.jboss.reddeer.examples/pom.xml
@@ -1,0 +1,15 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+
+	<modelVersion>4.0.0</modelVersion>
+	<artifactId>org.jboss.reddeer.examples</artifactId>
+	<packaging>eclipse-plugin</packaging>
+	<name>Red Deer Examples</name>
+	
+	<parent>
+		<groupId>org.jboss.reddeer</groupId>
+		<artifactId>parent</artifactId>
+		<version>0.2.0-SNAPSHOT</version>
+		<relativePath>../org.jboss.reddeer.parent/pom.xml</relativePath>
+	</parent>
+</project>
