+++ org.jboss.reddeer.examples/src/org/jboss/reddeer/junit/configuration/advanced/user-schema.xsd
