+++ org.jboss.reddeer.examples/src/org/jboss/reddeer/junit/configuration/simple/RedDeerRequirements.xsd
@@ -1,0 +1,74 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!-- RedDeer by Red Hat -->
+<xs:schema xmlns:xs="http://www.w3.org/2001/XMLSchema"
+	targetNamespace="http://www.jboss.org/NS/Req" xmlns:rd="http://www.jboss.org/NS/Req"
+	elementFormDefault="qualified">
+
+	<!-- Root element -->
+	<xs:element name="testrun" type="rd:testrunType">
+		<!-- Unique requirement 'name' attribute restriction -->
+		<xs:unique name="uniqueReqName">
+			<xs:selector xpath="*/*" />
+			<xs:field xpath="@name" />
+		</xs:unique>
+	</xs:element>
+
+	<!-- Test Run -->
+	<xs:complexType name="testrunType">
+		<xs:sequence>
+			<xs:element name="requirements" type="rd:requirementsType"
+				maxOccurs="1" />
+		</xs:sequence>
+	</xs:complexType>
+
+	<!-- Requirements -->
+	<xs:complexType name="requirementsType">
+		<xs:choice>
+			<xs:element ref="rd:abstractRequirement" minOccurs="1"
+				maxOccurs="unbounded" />
+		</xs:choice>
+	</xs:complexType>
+
+	<xs:element name="abstractRequirement" abstract="true"
+		type="rd:abstractRequirementType" />
+
+	<xs:complexType name="abstractRequirementType">
+		<xs:attribute name="name" type="xs:string" use="required" />
+	</xs:complexType>
+
+	<xs:element name="requirement" type="rd:requirementType"
+		substitutionGroup="rd:abstractRequirement">
+		<xs:annotation>
+			<xs:documentation>Specifies basic requirement structure - attribute
+				name needs to be unique in whole document. Class which fulfills the
+				Requirements needs to be set, additional properties can be set by
+				property elements.
+			</xs:documentation>
+		</xs:annotation>
+	</xs:element>
+
+	<!-- Requirement -->
+	<xs:complexType name="requirementType">
+		<xs:complexContent>
+			<xs:extension base="rd:abstractRequirementType">
+				<xs:sequence>
+					<xs:element name="property" type="rd:propertyType"
+						minOccurs="0" maxOccurs="unbounded">
+						<xs:annotation>
+							<xs:documentation>Well-known key-value properties.
+							</xs:documentation>
+						</xs:annotation>
+					</xs:element>
+				</xs:sequence>
+				<xs:attribute name="class" type="xs:string" use="required" />
+			</xs:extension>
+		</xs:complexContent>
+	</xs:complexType>
+
+	<!-- Property -->
+	<xs:complexType name="propertyType">
+		<xs:attribute name="key" type="xs:string" use="required" />
+		<xs:attribute name="value" type="xs:string" use="required" />
+	</xs:complexType>
+
+</xs:schema>
