+++ org.jboss.reddeer.examples/src/org/jboss/reddeer/junit/injection/RedDeerRequirements.xsd
