+++ org.jboss.reddeer.examples/src/org/jboss/reddeer/junit/injection/UserRequirement.java
@@ -1,0 +1,55 @@
+package org.jboss.reddeer.junit.injection;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+import org.jboss.reddeer.junit.requirement.PropertyConfiguration;
+import org.jboss.reddeer.junit.requirement.Requirement;
+import org.jboss.reddeer.junit.injection.UserRequirement.User;
+
+public class UserRequirement implements Requirement<User>, PropertyConfiguration{
+
+	@Retention(RetentionPolicy.RUNTIME)
+	@Target(ElementType.TYPE)
+	public @interface User {
+	}
+	
+	private String name;
+	
+	private String ip;
+	
+	private String password;
+	
+	public boolean canFulfill() {
+		// return true if you can connect to the database
+		return true;
+	}
+
+	public void fulfill() {
+		System.out.println("Fulfilling User requirement with\nName: " + name + "\nIP: " + ip);
+		// create an admin user in the database if it does not exist yet
+	}
+	
+	@Override
+	public void setDeclaration(User user) {
+		// annotation has no parameters no need to store reference to it
+	}
+	
+	public void setName(String name) {
+		this.name = name;
+	}
+	
+	public void setIp(String ip) {
+		this.ip = ip;
+	}
+	
+	public void setPassword(String password) {
+		this.password = password;
+	}
+	
+	public String getPassword() {
+		return password;
+	}
+}
