+++ org.jboss.reddeer.examples/src/org/jboss/reddeer/junit/injection/requirements.xml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="UTF-8"?>
+  <testrun
+        xmlns="http://www.jboss.org/NS/Req"
+        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+        xsi:schemaLocation="http://www.jboss.org/NS/Req RedDeerRequirements.xsd">
+
+	<requirements>
+		<requirement class="org.jboss.reddeer.junit.injection.UserRequirement" name="userRequirement">
+			<property key="name" value="USERS_ADMINISTRATION" />
+			<property key="ip" value="127.0.0.1" />
+			<property key="password" value="abc123" />
+		</requirement>
+	</requirements>
+</testrun>
