+++ org.jboss.reddeer.parent/pom.xml
@@ -33,7 +33,7 @@
 		<module>../org.jboss.reddeer.junit</module>
 		<module>../org.jboss.reddeer.junit.test</module>
 		<module>../org.jboss.reddeer.selenium</module>
-		<module>../org.jboss.reddeer.wiki.examples</module>
+		<module>../org.jboss.reddeer.examples</module>
 		<module>../org.jboss.reddeer.site</module>
 		<module>../org.jboss.reddeer.requirements</module>
 		<module>../org.jboss.reddeer.requirements.test</module>
