--- org.jboss.reddeer.wiki.examples/src/org/jboss/reddeer/wiki/introduction/example/HelloWorld.java
@@ -1,16 +1,0 @@
-package org.jboss.reddeer.wiki.introduction.example;
-
-import org.jboss.reddeer.eclipse.wst.server.ui.view.Server;
-import org.jboss.reddeer.eclipse.wst.server.ui.view.ServersView;
-
-public class HelloWorld {
-
-	public HelloWorld() {
-	}
-
-	public static void main(String[] args) {
-		ServersView serversView = new ServersView();
-		Server server = serversView.getServer("My Server");
-		server.start();
-	}
-}
