--- org.jboss.reddeer.wiki.examples/src/org/jboss/reddeer/wiki/junit/configuration/simple/requirements.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-  <testrun
-        xmlns="http://www.jboss.org/NS/Req"
-        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-        xsi:schemaLocation="http://www.jboss.org/NS/Req RedDeerSchema.xsd">
-
-	<requirements>
-		<requirement class="org.jboss.reddeer.wiki.user.db.simple.UserRequirement" name="userRequirement">
-			<property key="name" value="USERS_ADMINISTRATION" />
-			<property key="ip" value="127.0.0.1" />
-		</requirement>
-	</requirements>
-</testrun>
